fix compile warnings
vender and product id now have a type should be 32bit accourting to the usb standard

+++ inc/USB/common/FSUSBJoystickDeviceManager.h
@@ -39,7 +39,7 @@ namespace freestick
         virtual void init( );
         const FSUSBJoystick * getUSBJoystickDevice(unsigned int deviceID){return static_cast<const FSUSBJoystick *>(FSUSBDeviceManager::getDevice(deviceID));}
         virtual void updateEvents(unsigned int joystickDeviceID,FSUSBJoyStickInputElement * elementDevice, PhysicalValueNumber elementValue);
-        static long int createIdForElement(long int usage, long int usagePage);
+        static uint32_t createIdForElement(uint32_t usage, uint32_t usagePage);
 
     };
 }
